+++ android/src/main/java/com/applovin/reactnative/AppLovinMAXNativeAdView.java
@@ -391,6 +391,11 @@ public class AppLovinMAXNativeAdView
         nativeAdInfo.putString( "body", ad.getBody() );
         nativeAdInfo.putString( "callToAction", ad.getCallToAction() );
 
+        if ( ad.getStarRating() != null )
+        {
+            nativeAdInfo.putDouble( "starRating", ad.getStarRating().doubleValue() );
+        }
+
         float aspectRatio = ad.getMediaContentAspectRatio();
         if ( !Float.isNaN( aspectRatio ) )
         {
@@ -423,6 +428,11 @@ public class AppLovinMAXNativeAdView
         jsNativeAd.putString( "advertiser", ad.getAdvertiser() );
         jsNativeAd.putString( "body", ad.getBody() );
         jsNativeAd.putString( "callToAction", ad.getCallToAction() );
+
+        if ( ad.getStarRating() != null )
+        {
+            jsNativeAd.putDouble( "starRating", ad.getStarRating().doubleValue() );
+        }
 
         MaxNativeAdImage icon = ad.getIcon();
         if ( icon != null )